+++ warthog/packages/__init__.py
@@ -1,0 +1,15 @@
+# -*- coding: utf-8 -*-
+#
+# Warthog - Simple client for A10 load balancers
+#
+# Copyright 2014 Smarter Travel
+#
+# Available under the MIT license. See LICENSE for details.
+#
+
+"""
+warthog.packages
+~~~~~~~~~~~~~~~~
+
+Third-party libs that we bundle with Warthog
+"""